test(WebAPI): Fixing unit tests
